@@ -8,8 +8,10 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "bootstrap": "^5.3.2",
+    "bootstrap-vue-3": "^0.5.1",
     "core-js": "^3.8.3",
-    "vue": "^3.2.13"
+    "vue": "^3.3.7"
   },
   "devDependencies": {
     "@babel/core": "^7.12.16",
