@@ -1,26 +1,20 @@
 <template>
-  <img alt="Vue logo" src="./assets/logo.png">
-  <HelloWorld msg="Welcome to Your Vue.js App"/>
+    <body>
+        <div>
+            <b-button>Button</b-button>
+            <b-button variant="danger">Button</b-button>
+            <b-button variant="success">Button</b-button>
+            <b-button variant="outline-primary">Button</b-button>
+        </div>
+    </body>
 </template>
 
-<script>
-import HelloWorld from './components/HelloWorld.vue'
-
-export default {
-  name: 'App',
-  components: {
-    HelloWorld
-  }
-}
-</script>
+<script></script>
 
 <style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 60px;
-}
+    body {
+        background-color: black;
+        color: white;
+        /* 텍스트 색상을 흰색으로 설정 */
+    }
 </style>
